processor/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
email-validator>=2.1.0
python-magic>=0.4.27
python-dotenv>=1.0.0